EES-5133 Refactor `DataSetControllerGetQueryTests` to run faster
